test(ertp-ledgerguise): add adversarial suite
- move attack tests into dedicated file

- assert commodity GUID collisions are rejected

- fix createIssuerKit to fail on existing commodity
